when adding new panel try to find a spot that is empty, add conversion or
Sat Dec 19 21:10:02 1998  George Lebl  <jirka@5z.com>

        * menu.c: when adding new panel try to find a spot that is empty,
          add conversion or edge<->corner panel and move add drawer to
          the main panel menu from the add panel submenu

        * panel.c: don't put up panel menu on the left click, just on the
          right click

        * panel_config.c,panel_config_global.c,menu.c: capitalize only the
          first letter in strings for consistency

+++ gnome-panel/extern.c
@@ -402,7 +402,7 @@ s_panel_add_applet_full(POA_GNOME_Panel *servant,
 				*globcfgpath = CORBA_string_dup(old_panel_cfg_path);
 				info->type = APPLET_EXTERN_RESERVED;
 				*wid=GDK_WINDOW_XWINDOW(socket->window);
-				printf("\nSOCKET XID: %lX\n\n", *wid);
+				printf("\nSOCKET XID: %lX\n\n", (long)*wid);
 
 				panelspot_servant = (POA_GNOME_PanelSpot *)ext;
 				acc = PortableServer_POA_servant_to_reference(thepoa, panelspot_servant, ev);
